requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
python-dotenv==1.0.0
python-http-client==3.3.7
requests==2.28.2
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
urllib3==1.26.14
